#526 Add dossier metric integration tests
